chore: move BatchFileUploader to the root of the packages
